requirements.txt
pypiwin32==223
tensorflow-gpu==1.9.0
tensorlayer==1.8.0
easydict==1.9
numpy==1.16.1